requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
blessed==1.20.0
certifi
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
docker-pycreds==0.4.0
emoji==0.6.0
entrypoints==0.4
filelock==3.10.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
gensim==3.8.1
gitdb==4.0.10
GitPython==3.1.31
gpustat==1.1
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.5
jedi==0.18.2
jieba==0.42.1
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.6
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.21.6
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==11.525.112
openai==0.27.4
packaging==23.0
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
preprocess==2.0.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
regex==2022.10.31
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.17.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
some-package==0.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
tweet-preprocessor==0.6.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
wcwidth==0.2.6
widgetsnbextension==4.0.6
wordninja==2.0.0
yarl==1.8.2
zipp==3.15.0
